+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - released beta version of improved polling performance
 - added `yamllint` validation for the `values.yaml` formatting
 - added "in code" validation of groups and profiles
+- added logs configuration to docker compose deployment
 
 ### Fixed
 - fixed a bug with configuration from values.yaml not being transferred to the UI while migrating to SC4SNMP-UI